Preferences UI: add dive log tab
This adds a tab for dive log - related preferences.
A suitable test programs is still required.
