+++ core/settings/qPref.cpp
@@ -16,6 +16,7 @@
 #include "qPrefUpdateManager.h"
 #include "qPrefEquipment.h"
 #include "qPrefMedia.h"
+#include "qPrefLog.h"
 
 #include <QtQml>
 #include <QQmlContext>
@@ -42,6 +43,7 @@ void qPref::loadSync(bool doSync)
 	qPrefUpdateManager::loadSync(doSync);
 	qPrefEquipment::loadSync(doSync);
 	qPrefMedia::loadSync(doSync);
+	qPrefLog::loadSync(doSync);
 }
 
 Q_DECLARE_METATYPE(deco_mode);
@@ -68,6 +70,7 @@ void qPref::registerQML(QQmlEngine *engine)
 		ct->setContextProperty("PrefEquipment", qPrefUpdateManager::instance());
 		ct->setContextProperty("PrefMedia", qPrefUpdateManager::instance());
 		ct->setContextProperty("PrefClearDc", qPrefUpdateManager::instance());
+		ct->setContextProperty("PrefLog", qPrefUpdateManager::instance());
 	}
 
 	// Register special types
