+++ desktop-widgets/preferences/preferences_defaults.ui
@@ -46,101 +46,7 @@
      </layout>
     </widget>
    </item>
-   <item>
-    <widget class="QGroupBox" name="groupBox_3">
-     <property name="title">
-      <string>Default file</string>
-     </property>
-     <layout class="QFormLayout" name="formLayout_2">
-      <property name="horizontalSpacing">
-       <number>5</number>
-      </property>
-      <property name="verticalSpacing">
-       <number>5</number>
-      </property>
-      <property name="margin">
-       <number>5</number>
-      </property>
-      <item row="0" column="0">
-       <widget class="QLabel" name="defaultDiveLogFileLabel">
-        <property name="text">
-         <string>Default dive log file</string>
-        </property>
-       </widget>
-      </item>
-      <item row="0" column="1">
-       <layout class="QHBoxLayout" name="defaultFileBehaviorLayout">
-        <item>
-         <widget class="QRadioButton" name="noDefaultFile">
-          <property name="text">
-           <string>&amp;No default file</string>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <widget class="QRadioButton" name="localDefaultFile">
-          <property name="text">
-           <string>&amp;Local default file</string>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <widget class="QRadioButton" name="cloudDefaultFile">
-          <property name="text">
-           <string>Clo&amp;ud storage default file</string>
-          </property>
-         </widget>
-        </item>
-       </layout>
-      </item>
-      <item row="1" column="0">
-       <widget class="QLabel" name="label_9">
-        <property name="text">
-         <string>Local dive log file</string>
-        </property>
-       </widget>
-      </item>
-      <item row="1" column="1">
-       <layout class="QHBoxLayout" name="horizontalLayout_2">
-        <item>
-         <widget class="QLineEdit" name="defaultfilename"/>
-        </item>
-        <item>
-         <widget class="QToolButton" name="btnUseDefaultFile">
-          <property name="text">
-           <string>Use default</string>
-          </property>
-          <property name="checkable">
-           <bool>true</bool>
-          </property>
-         </widget>
-        </item>
-        <item>
-         <widget class="QToolButton" name="chooseFile">
-          <property name="text">
-           <string>...</string>
-          </property>
-         </widget>
-        </item>
-       </layout>
-      </item>
-      <item row="2" column="0">
-       <widget class="QLabel" name="label_10">
-        <property name="text">
-         <string>Display invalid</string>
-        </property>
-       </widget>
-      </item>
-      <item row="2" column="1">
-       <widget class="QCheckBox" name="displayinvalid">
-        <property name="text">
-         <string/>
-        </property>
-       </widget>
-      </item>
-     </layout>
-    </widget>
-   </item>
+
    <item>
     <widget class="QGroupBox" name="groupBox_7">
      <property name="title">
@@ -177,6 +83,7 @@
      </layout>
     </widget>
    </item>
+
    <item>
     <widget class="QPushButton" name="resetSettings">
      <property name="text">
@@ -184,22 +91,7 @@
      </property>
     </widget>
    </item>
-   <item>
-    <widget class="QGroupBox" name="groupBox_extra_star_widgets">
-     <property name="title">
-      <string>Environmental parameters in Information tab</string>
-     </property>
-     <layout class="QHBoxLayout" name="extraWidgetLayout">
-      <item>
-       <widget class="QCheckBox" name="extraEnvironmentalDefault">
-        <property name="text">
-         <string>Show extended range of environmental paramaters</string>
-        </property>
-       </widget>
-      </item>
-     </layout>
-    </widget>
-   </item>
+
    <item>
     <spacer name="verticalSpacer_2">
      <property name="orientation">
