+++ desktop-widgets/preferences/preferences_equipment.cpp
@@ -11,7 +11,7 @@
 
 #include "qt-models/models.h"
 
-PreferencesEquipment::PreferencesEquipment() : AbstractPreferencesWidget(tr("Equipment"), QIcon(":preferences-equipment-icon"), 4)
+PreferencesEquipment::PreferencesEquipment() : AbstractPreferencesWidget(tr("Equipment"), QIcon(":preferences-equipment-icon"), 5)
 {
 	ui = new Ui::PreferencesEquipment();
 	ui->setupUi(this);
