+++ desktop-widgets/preferences/preferences_georeference.cpp
@@ -7,7 +7,7 @@
 
 #include <ctime>
 
-PreferencesGeoreference::PreferencesGeoreference() : AbstractPreferencesWidget(tr("Georeference"), QIcon(":geotag-icon"), 9)
+PreferencesGeoreference::PreferencesGeoreference() : AbstractPreferencesWidget(tr("Georeference"), QIcon(":geotag-icon"), 8)
 {
 	ui = new Ui::PreferencesGeoreference();
 	ui->setupUi(this);
