+++ desktop-widgets/preferences/preferences_log.ui
@@ -1,0 +1,199 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<ui version="4.0">
+ <class>PreferencesLog</class>
+ <widget class="QWidget" name="PreferencesLog">
+  <property name="geometry">
+   <rect>
+    <x>0</x>
+    <y>0</y>
+    <width>561</width>
+    <height>558</height>
+   </rect>
+  </property>
+  <property name="windowTitle">
+   <string>Form</string>
+  </property>
+  <layout class="QVBoxLayout" name="verticalLayout">
+
+      <item>
+       <widget class="QLabel" name="label_9">
+        <property name="text">
+         <string>DIVE LOG</string>
+        </property>
+       </widget>
+      </item>
+
+   <item>
+    <widget class="QGroupBox" name="groupBox_3">
+     <property name="title">
+      <string></string>
+     </property>
+     <layout class="QFormLayout" name="formLayout_2">
+      <property name="horizontalSpacing">
+       <number>5</number>
+      </property>
+      <property name="verticalSpacing">
+       <number>5</number>
+      </property>
+      <property name="margin">
+       <number>5</number>
+      </property>
+      <item row="0" column="0">
+       <widget class="QLabel" name="defaultDiveLogFileLabel">
+        <property name="text">
+         <string>Default dive log file</string>
+        </property>
+       </widget>
+      </item>
+      <item row="0" column="1">
+       <layout class="QHBoxLayout" name="defaultFileBehaviorLayout">
+        <item>
+         <widget class="QRadioButton" name="noDefaultFile">
+          <property name="text">
+           <string>&amp;No default file</string>
+          </property>
+         </widget>
+        </item>
+        <item>
+         <widget class="QRadioButton" name="localDefaultFile">
+          <property name="text">
+           <string>&amp;Local default file</string>
+          </property>
+         </widget>
+        </item>
+        <item>
+         <widget class="QRadioButton" name="cloudDefaultFile">
+          <property name="text">
+           <string>Clo&amp;ud storage default file</string>
+          </property>
+         </widget>
+        </item>
+       </layout>
+      </item>
+      <item row="1" column="0">
+       <widget class="QLabel" name="label_9">
+        <property name="text">
+         <string>Local dive log file</string>
+        </property>
+       </widget>
+      </item>
+      <item row="1" column="1">
+       <layout class="QHBoxLayout" name="horizontalLayout_2">
+        <item>
+         <widget class="QLineEdit" name="defaultfilename"/>
+        </item>
+        <item>
+         <widget class="QToolButton" name="btnUseDefaultFile">
+          <property name="text">
+           <string>Use default</string>
+          </property>
+          <property name="checkable">
+           <bool>true</bool>
+          </property>
+         </widget>
+        </item>
+        <item>
+         <widget class="QToolButton" name="chooseFile">
+          <property name="text">
+           <string>...</string>
+          </property>
+         </widget>
+        </item>
+       </layout>
+      </item>
+     </layout>
+    </widget>
+   </item>
+
+   <item>
+    <widget class="QCheckBox" name="displayinvalid">
+     <property name="text">
+      <string>In the Dive List, show dives that you have marked as invalid, e.g. pool dives</string>
+     </property>
+    </widget>
+   </item>
+
+   <item>
+    <widget class="QCheckBox" name="extraEnvironmentalDefault">
+     <property name="text">
+      <string>In the Information tab, show tools for recording wave height, surge and chill during a dive</string>
+     </property>
+    </widget>
+   </item>
+
+   <item>
+    <widget class="QCheckBox" name="show_average_depth">
+     <property name="text">
+      <string>In the Dive Profile, show mean depth</string>
+     </property>
+    </widget>
+   </item>
+
+   <item>
+    <spacer name="verticalSpacer_2">
+     <property name="orientation">
+      <enum>Qt::Vertical</enum>
+     </property>
+     <property name="sizeHint" stdset="0">
+      <size>
+       <width>0</width>
+       <height>195</height>
+      </size>
+     </property>
+    </spacer>
+   </item>
+  </layout>
+ </widget>
+
+ <resources/>
+ <connections>
+  <connection>
+   <sender>btnUseDefaultFile</sender>
+   <signal>toggled(bool)</signal>
+   <receiver>chooseFile</receiver>
+   <slot>setHidden(bool)</slot>
+   <hints>
+    <hint type="sourcelabel">
+     <x>462</x>
+     <y>136</y>
+    </hint>
+    <hint type="destinationlabel">
+     <x>528</x>
+     <y>145</y>
+    </hint>
+   </hints>
+  </connection>
+  <connection>
+   <sender>velocitySlider</sender>
+   <signal>valueChanged(int)</signal>
+   <receiver>velocitySpinBox</receiver>
+   <slot>setValue(int)</slot>
+   <hints>
+    <hint type="sourcelabel">
+     <x>437</x>
+     <y>299</y>
+    </hint>
+    <hint type="destinationlabel">
+     <x>531</x>
+     <y>301</y>
+    </hint>
+   </hints>
+  </connection>
+  <connection>
+   <sender>velocitySpinBox</sender>
+   <signal>valueChanged(int)</signal>
+   <receiver>velocitySlider</receiver>
+   <slot>setValue(int)</slot>
+   <hints>
+    <hint type="sourcelabel">
+     <x>547</x>
+     <y>274</y>
+    </hint>
+    <hint type="destinationlabel">
+     <x>482</x>
+     <y>276</y>
+    </hint>
+   </hints>
+  </connection>
+ </connections>
+</ui>
