+++ desktop-widgets/preferences/preferences_network.cpp
@@ -6,7 +6,7 @@
 #include "core/settings/qPrefProxy.h"
 #include <QNetworkProxy>
 
-PreferencesNetwork::PreferencesNetwork() : AbstractPreferencesWidget(tr("Network"),QIcon(":preferences-system-network-icon"), 9), ui(new Ui::PreferencesNetwork())
+PreferencesNetwork::PreferencesNetwork() : AbstractPreferencesWidget(tr("Network"),QIcon(":preferences-system-network-icon"), 10), ui(new Ui::PreferencesNetwork())
 {
 	ui->setupUi(this);
